Move testing related function to a dedicated lib + move unit tests
