+++ tasks/testwasher.py
@@ -11,6 +11,7 @@ from tasks.libs.ciproviders.gitlab_api import (
     resolve_gitlab_ci_configuration,
 )
 from tasks.libs.common.utils import gitlab_section
+from tasks.libs.testing.flakes import get_tests_family_if_failing_tests, is_known_flaky_test
 from tasks.test_core import ModuleTestResult
 
 FLAKY_TEST_INDICATOR = "flakytest: this is a known flaky test"
@@ -40,11 +41,11 @@ class TestWasher:
 
         for package, tests in failing_tests.items():
             non_flaky_failing_tests_in_package = set()
-            known_flaky_tests_parents = self.get_tests_family_if_failing_tests(
+            known_flaky_tests_parents = get_tests_family_if_failing_tests(
                 all_known_flakes[package], failing_tests[package]
             )
             for failing_test in tests:
-                if not self.is_known_flaky_test(failing_test, all_known_flakes[package], known_flaky_tests_parents):
+                if not is_known_flaky_test(failing_test, all_known_flakes[package], known_flaky_tests_parents):
                     non_flaky_failing_tests_in_package.add(failing_test)
             if non_flaky_failing_tests_in_package:
                 non_flaky_failing_tests[package] = non_flaky_failing_tests_in_package
@@ -124,53 +125,6 @@ class TestWasher:
             print("Please check the job logs for more information")
 
         return should_succeed
-
-    def is_known_flaky_test(self, failing_test, known_flaky_tests, known_flaky_tests_parents):
-        """
-        Check if a test is known to be flaky
-        The method should be called with the following arguments:
-        - failing_test: the test that is failing
-        - known_flaky_tests: the set of tests that are known to be flaky
-        - known_flaky_tests_parents: the set of tests that are ancestors of a known flaky test, thus would fail when the flaky leaf test fails
-        If a test is a parent of a test that is known to be flaky, the test should be considered flaky
-        For example:
-        - if TestEKSSuite/TestCPU is known to be flaky, TestEKSSuite/TestCPU/TestCPUUtilization should be considered flaky
-        - if TestEKSSuite/TestCPU is known to be flaky, TestEKSSuite should be considered flaky unless TestEKSSuite/TestCPU is not failing
-        - if TestEKSSuite/TestCPU is known to be flaky, TestEKSSuite/TestMemory should not be considered flaky
-        """
-
-        failing_test_parents = self.get_tests_family([failing_test])
-
-        if any(parent in known_flaky_tests for parent in failing_test_parents):
-            return True
-
-        return failing_test in known_flaky_tests_parents
-
-    def get_tests_family_if_failing_tests(self, test_name_list, failing_tests: set):
-        """
-        Get the parent tests of a list of tests only if the marked test is failing
-        For example with the test ["TestEKSSuite/TestCPU/TestCPUUtilization", "TestKindSuite/TestCPU"]
-        this method should return the set{"TestEKSSuite/TestCPU/TestCPUUtilization", "TestEKSSuite/TestCPU", "TestEKSSuite", "TestKindSuite/TestCPU", "TestKindSuite"}
-        if TestKindSuite/TestCPU and TestEKSSuite/TestCPU/TestCPUUtilization are failing
-        Another example, with the test ["TestEKSSuite/TestCPU/TestCPUUtilization", "TestKindSuite/TestCPU"]
-        if only TestKindSuite/TestCPU is failing, the method should return the set{"TestKindSuite/TestCPU", "TestKindSuite"}
-        """
-        test_name_set = set(test_name_list)
-        marked_tests_failing = failing_tests.intersection(test_name_set)
-        return self.get_tests_family(list(marked_tests_failing))
-
-    def get_tests_family(self, test_name_list):
-        """
-        Get the parent tests of a list of tests
-        For example with the test ["TestEKSSuite/TestCPU/TestCPUUtilization", "TestKindSuite/TestCPU"]
-        this method should return the set{"TestEKSSuite/TestCPU/TestCPUUtilization", "TestEKSSuite/TestCPU", "TestEKSSuite", "TestKindSuite/TestCPU", "TestKindSuite"}
-        """
-        test_family = set(test_name_list)
-        for test_name in test_name_list:
-            while test_name.count('/') > 0:
-                test_name = test_name.rsplit('/', 1)[0]
-                test_family.add(test_name)
-        return test_family
 
 
 @task
